Clarify that the .well-known method for Jitsi isn't available yet
We should remove this at release time, but if we forget it should be vaguely worded well enough.

People keep assuming it'll just work even though the feature hasn't launched yet. The original theory was that a release was around the corner, but it's since been pushed back several times.